Vision data-throughput fixes: 64 workers, OMP=1, bf16 inputs, scale-gated grad ckpt
GPU was data-starved (~40% util, MFU 5-7%): 16 decode workers of 72 cores.
- num_workers 16->64, OMP_NUM_THREADS 32->1 (avoid worker thread oversubscription)
- ship bf16 input tensors (halves pinned prefetch buffers; first layer autocasts anyway)
- grad checkpointing only s4/s5 (smaller scales fit, drop ~30% recompute)

+++ experiments/vision/hf_vision_data.py
@@ -62,7 +62,7 @@ class HFVisionDataset(Dataset):
         sample = self._hf[idx]
         image = sample[self._preset.image_key].convert("RGB")
         label = sample[self._preset.label_key]
-        x = self._transform(image)
+        x = self._transform(image).to(torch.bfloat16)
         y = torch.tensor(label, dtype=torch.long)
         return x, y
